Shorter ISA bus delays (using a for-loop - should probably use delayMicroseconds but I'm trying to reduce waits to as little as possible).
Added some documentation on OPL-3 matters.

canyon.ino has mostly experimental code in it at the moment, which can handle a LOT
of incoming MIDI data on the MPU-401. Interrupts are used (on pin 2) with a circular
buffer.

Some prototype code has been added, which runs on Windows and basically sends a
custom SysEx message (F0 7D xa 0b 0c 0d F7) where x is 0/1 for primary/secondary
register set, a/b are the high/low nybbles of an OPL-3 register, and c/d are the
high/low nybbles of the data byte. This allows OPL-3 control to take place via a
PC for easier prototyping of the synth code.