Actually calculate each note's frequency instead of just using a lookup table.
This would involve some floating-point calculations so the frequencies are
specified in 1/100th Hz to deal with fractions easily. Also we'd then end up
with huge integers that exceed 32 bits...

Now there is a simple check to determine the block number based on the
frequency, a table of "block multipliers" which determines how much to
multiply the frequency before dividing the result by 2485800 (which is
49716 * 200, because the multiplier is half what it should be to keep it
in range of a uint32_t).

A table of note frequencies is used still, but this is specified as 1/100th
Hz instead of f-numbers.

It's all very complicated but avoids using floating point numbers and should
save us a few calculations along the way too.