Create OPL3Synth and OPL3Operator classes. These abstract the channels/operators offered by OPL3 into a convenient, intuitive API.
Channels can be allocated/freed. It is up to the caller to determine when to free
a channel (i.e. if a long sustain has been set, it is desirable to let the release
phase of the envelope complete).

There needs to be a way to enable percussion mode and allocate 4-op channels. Again,
this will be the responsibility of the caller - if it cannot allocate a 4-op channel,
it should identify (somehow) which channels it is using that are capable of doing
4-op and forcibly free them.

Essentially this just manages the OPL3 resources, according to whatever the caller
requests.

canyon.ino has been updated to use the new methods.