Re-enable use of interrupts because programming multiple OPL3 parameters can take a few milliseconds (can probably reduce this by not committing every individual change - e.g. setting attack/decay separately causes two writes to the same register so we could cache it).
The delay was causing MPU-401 buffer to fill, resulting in stuck notes.

Also added ISRState as a means of determining whether interrupts should be disabled and
re-enabled during the ISABus transactions - if we enable it during an interrupt handler,
this meant the completion of a read inside the interrupt handler would immediately
trigger the interrupt again so we'd end up somehow fetching data in reverse order
(because of the order in which the handlers returned).