Turn on diagnostic LED on pin 10 during startup. After startup completes, turn the LED off.
If there is a problem during startup, the LED will flash a number of times (varies depending
on the problem) before pausing for a couple of seconds and repeating the same number of
flashes.

Otherwise, the LED turns off once startup completes. It will then flash as MIDI activity
is detected. It is connected to a PWM pin so there is some code to fade the LED whilst there
is no activity. This makes it easier to spot when there is activity.